Rung 405 rerun audit: aggregation repair moved wrong way (8.6e-4); equivalence robust across both; shared scoring.py proposed

+++ basis_aligned/bilinear_quotient/ops/mlp0_rank448_token_grammar_active_subspace.py
@@ -246,7 +246,8 @@ def _score_wave(model, rows, programs, rung403, device):
         logits = _manual_logits(model, index)
         handle.remove()
         losses = F.cross_entropy(
-            logits[:, SCORING].transpose(1, 2), target[:, SCORING], reduction="none"
+            logits[:, SCORING].float().transpose(1, 2),
+            target[:, SCORING], reduction="none"
         ).mean(1)
         document_ce["native"].extend(float(loss) for loss in losses)
         for name, program in programs.items():
@@ -264,7 +265,8 @@ def _score_wave(model, rows, programs, rung403, device):
             logits = _manual_logits(model, index)
             handle.remove()
             losses = F.cross_entropy(
-                logits[:, SCORING].transpose(1, 2), target[:, SCORING], reduction="none"
+                logits[:, SCORING].float().transpose(1, 2),
+                target[:, SCORING], reduction="none"
             ).mean(1)
             document_ce[name].extend(float(loss) for loss in losses)
     ce = {
